# docs/api-docs.yaml
openapi: 3.0.1
info:
  title: Bank
  description: Тестовое задание
  contact:
    name: Ветрова Анастасия
    email: Vetrova600@gmail.com
  version: v1
servers:
- url: https://localhost:8443
  description: Generated server url
security:
- bearerAuth: []
paths:
  /api/v1/cards:
    get:
      tags:
      - Карты
      summary: Получение списка карт текущего пользователя (USER)
      description: |
        Возвращает список карт, принадлежащих авторизованному пользователю.
        Доступно только для пользователей с ролью USER.
        Информация выдается постранично.
      operationId: getUserCards
      parameters:
      - name: page
        in: query
        required: false
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 1
      - name: size
        in: query
        required: false
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: Список карт успешно получен
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserCardsResponse"
        "403":
          description: Нет доступа — пользователь не авторизован или недостаточно
            прав
          content:
            text/plain:
              schema:
                type: string
    post:
      tags:
      - Карты
      summary: Создание новой банковской карты (ADMIN)
      description: |
        Позволяет администратору создать карту для пользователя.
        Карта будет активна с начальным балансом и сроком действия в годах.
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardDto"
        required: true
      responses:
        "201":
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        "400":
          description: Ошибка валидации входных данных
          content:
            text/plain:
              schema:
                type: string
        "403":
          description: Недостаточно прав доступа
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: Пользователь не найден
          content:
            text/plain:
              schema:
                type: string
  /api/v1/cards/transfer:
    post:
      tags:
      - Карты
      summary: Перевод средств между картами пользователя (USER)
      description: |
        Позволяет пользователю перевести средства между своими банковскими картами
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
        required: true
      responses:
        "200":
          description: Перевод выполнен успешно
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "400":
          description: Недостаточно средств или некорректные данные
          content:
            text/plain:
              schema:
                type: string
        "403":
          description: Доступ запрещён
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: Одна из карт не найдена
          content:
            text/plain:
              schema:
                type: string
        "409":
          description: Недостаточно средств
          content:
            text/plain:
              schema:
                type: string
...
  /api/v1/cards/status:
    patch:
      tags:
      - Карты
      summary: Обновление статуса карты (ADMIN)
      description: |
        Позволяет администратору изменить статус карты.
        Передавайте `userId`, `cardId` и новый `cardStatus`(принимает значения "ACTIVE", "BLOCKED", "EXPIRED").
        Если карта уже находится в заданном статусе — возвращается ошибка.
      operationId: updateStatus
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardStatusRequest"
        required: true
      responses:
        "204":
          description: Статус карты успешно обновлён
        "400":
          description: Карта уже имеет указанный статус или переданы некорректные
            данные
        "403":
          description: Недостаточно прав для выполнения операции
        "404":
          description: Карта не найдена или не принадлежит пользователю
...
  /api/v1/cards/{cardId}/balance:
    get:
      tags:
      - Карты
      summary: Получить баланс карты (USER)
      description: |
        Позволяет пользователю получить текущий баланс своей банковской карты по её UUID
      operationId: getBalance
      parameters:
      - name: cardId
        in: path
        description: ID карты
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Баланс успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceResponse"
        "404":
          description: Карта не найдена или не принадлежит пользователю
          content:
            text/plain:
              schema:
                type: string
        "403":
          description: Недостаточно прав доступа
          content:
            text/plain:
              schema:
                type: string
  /api/v1/cards/all:
    get:
      tags:
      - Карты
      summary: Получение всех карт (ADMIN)
      description: |
        Возвращает постраничный список всех карт в системе.
        Доступно только для пользователей с ролью ADMIN.
        Поддерживается сортировка и пагинация.
      operationId: getAllCards
      parameters:
      - name: page
        in: query
        required: false
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 1
      - name: size
        in: query
        required: false
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 50
      - name: sortBy
        in: query
        description: сортировка по полю
        required: false
        schema:
          type: string
          enum:
          - user
          - cardNumber
          - expirationDate
          - status
          - balance
      - name: sortDir
        in: query
        required: false
        schema:
          type: string
          enum:
          - desc
          - asc
      responses:
        "200":
          description: Список карт успешно получен
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardResponse"
        "400":
          description: Некорректные параметры сортировки или пагинации
          content:
            text/plain:
              schema:
                type: string
        "403":
          description: Нет доступа — недостаточно прав
          content:
            text/plain:
              schema:
                type: string
  /api/v1/cards/{cardId}/users/{userId}:
    delete:
      tags:
      - Карты
      summary: Удалить карту пользователя (ADMIN)
      description: |
        Удаляет карту с указанным ID, если её баланс равен 0.
        Доступно только для администратора.
      operationId: deleteCard
      parameters:
      - name: cardId
        in: path
        description: ID карты
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        description: ID пользователя
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Карта успешно удалена
        "400":
          description: Баланс карты не равен 0
        "403":
          description: Недостаточно прав доступа
        "404":
          description: Карта не найдена или не принадлежит пользователю
components:
  schemas:
    CardResponse:
      type: object
      properties:
        id:
          type: string
          description: ID карты
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        userId:
          type: string
          description: ID владельца карты
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        cardNumber:
          type: string
          description: Номер карты (маскированный)
          example: '**** **** **** 1234'
        expirationDate:
          type: string
          description: Срок действия карты
          format: date
          example: 2025-12-31
        status:
          type: string
          description: Статус банковской карты
          example: ACTIVE
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        balance:
          type: number
          description: Баланс карты
          example: 1000.0
      description: Информация о карте
    CreateCardDto:
      required:
      - balance
      - userId
      - validityPeriodYears
      type: object
      properties:
        userId:
          type: string
          description: "ID пользователя, для которого создается карта"
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        validityPeriodYears:
          maximum: 5
          minimum: 1
          type: integer
          description: Срок действия карты в годах (от 1 до 5 лет)
          format: int32
          example: 3
        balance:
          minimum: 0.01
          exclusiveMinimum: false
          type: number
          description: Начальный баланс карты
          example: 1000.0
      description: DTO для создания новой банковской карты
    TransferResponse:
      type: object
      properties:
        fromCardId:
          type: string
          description: "ID карты, с которой был выполнен перевод"
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        fromCardBalanceAfter:
          type: number
          description: Оставшийся баланс на карте отправителя после перевода
          example: 1000.0
        toCardId:
          type: string
          description: "ID карты, на которую был выполнен перевод"
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        toCardBalanceAfter:
          type: number
          description: Баланс карты получателя после перевода
          example: 2500.5
      description: Ответ на успешный перевод средств между картами
    TransferRequest:
      required:
      - amount
      - fromCardId
      - toCardId
      type: object
      properties:
        fromCardId:
          type: string
          description: "ID карты, с которой осуществляется перевод"
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        toCardId:
          type: string
          description: "ID карты, на которую осуществляется перевод"
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        amount:
          maximum: 50000.00
          exclusiveMaximum: false
          minimum: 0.01
          exclusiveMinimum: false
          type: number
          description: Сумма перевода
          example: 999.99
      description: Запрос на перевод средств с одной карты на другую
...
    UpdateCardStatusRequest:
      required:
      - cardId
      - cardStatus
      - userId
      type: object
      properties:
        userId:
          type: string
          description: ID пользователя
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        cardId:
          type: string
          description: ID карты
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        cardStatus:
          type: string
          description: Статус банковской карты
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
      description: Запрос на обновление статуса карты
...
    UserCardsResponse:
      type: object
      properties:
        id:
          type: string
          description: ID карты
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        cardNumber:
          type: string
          description: Номер карты (маскированный)
          example: '**** **** **** 1234'
        expirationDate:
          type: string
          description: Дата окончания действия карты
          format: date
          example: 2025-12-31
        status:
          type: string
          description: Статус банковской карты
          example: ACTIVE
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
      description: Ответ с информацией по карте пользователя
    BalanceResponse:
      type: object
      properties:
        cardId:
          type: string
          description: ID карты
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        balance:
          type: number
          description: Баланс карты
          example: 1500.75
      description: Ответ с информацией о балансе карты
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT